Improve Orphan Prevention By Querying Space
The previous logic tried to match by title, but this isn't guaranteed
to be valid. Replace with more robust logic that queries the first page
of the space by the provided space key and uses that for orphan prevention.

This should be the root page of the space, and this should fix issue #5

Also update unit tests